renamed strain-config to strain-config-utils
