+++ src/strain-config-utils.js
